requirements.txt
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
cryptography==3.3.1
defusedxml==0.6.0
idna==2.10
oauthlib==3.1.0
pbr==5.5.1
pycparser==2.20
PyJWT==2.0.0
requests==2.25.1
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
six==1.15.0
todoist-python==8.1.3
urllib3==1.26.2
